Adds a button to be pushed at the end of the pause
Rather than assume the user is looking at the count down, and start to
use the computer immediately at the end of the pause, let him press a
button to exit the locked mode. That way, the countdown doesn't begin
while the user is away. The old behavior remains as an option, but is
off by default.  This solves #3846.